requirements.txt
trimesh[all]
numpy
pandas
tabulate
lpips
matplotlib
Pillow
torch 
torchvision
torchaudio
torchinfo
scikit-image
scikit-learn
imageio
tqdm
libigl
# pyembree
opencv-python
pyyaml
torchinfo==1.5.4
fast-simplification
pyarrow
gdown       # for downloading files from google drive. Needed for Diligent dataset
robust-laplacian
tensorboard
hydra-core
omegaconf
